Fixes #1469 by reverting 96186d6fd73f5b1e678ad0345696c002593e48b5.

I intend to do a more thorough look at how to get that performance back and redo the caching... This was just the fastest solution after writing the test and benchmark.

# Benchmark

```
BenchmarkDotNet v0.13.12, Windows 10 (10.0.19045.5247/22H2/2022Update)
AMD Ryzen 5 1600X, 1 CPU, 12 logical and 6 physical cores
.NET SDK 9.0.101
  [Host]     : .NET 6.0.36 (6.0.3624.51421), X64 RyuJIT AVX2
  DefaultJob : .NET 6.0.36 (6.0.3624.51421), X64 RyuJIT AVX2
```

## Before

| Method | RowCount | Mean     | Error     | StdDev    |
|------- |--------- |---------:|----------:|----------:|
| Read   | 1000     | 7.045 ms | 0.1364 ms | 0.1675 ms |

## After

| Method | RowCount | Mean     | Error     | StdDev    |
|------- |--------- |---------:|----------:|----------:|
| Read   | 1000     | 9.296 ms | 0.1820 ms | 0.2302 ms |

## Math

Looks to be a 32% regression in performance.  Pretty bad.