@Vemahk 2025-01-25
Added async local implementation & updated initial PR with new benchmarks